Courses view refractored

+++ app/service/examination_service.py
@@ -2,19 +2,23 @@
 
 from app.repository import examination_repository
 # from app.utils import file
-from app.views.errors import ResourceNotFoundException
+from app.exceptions import ResourceNotFoundException, BusinessRuleException, \
+    DuplicateResourceException
 
 
-def find_examination_by_id(examination_id):
+def get_examination_by_id(examination_id):
     return examination_repository.find_examination_by_id(examination_id)
 
 
-def find_education_by_id(education_id):
+def get_education_by_id(education_id):
     return examination_repository.find_education_by_id(education_id)
 
 
-def find_course_by_id(course_id):
-    return examination_repository.find_course_by_id(course_id)
+def get_course_by_id(course_id):
+    course = examination_repository.find_course_by_id(course_id)
+    if not course:
+        raise ResourceNotFoundException("Course not found", course_id)
+    return course
 
 
 def find_all_courses():
@@ -28,19 +32,31 @@ def find_all_educations():
 def find_all_examinations_by_course(course_id):
     exams = examination_repository.find_all_examinations_by_course(course_id)
     if not exams:
-        raise ResourceNotFoundException("Examination")
+        raise ResourceNotFoundException("Examinations")
     return exams
 
 
 def find_all_examinations_by_education(education_id):
     e = examination_repository.find_all_examinations_by_education(education_id)
     if not e:
-        raise ResourceNotFoundException("Examination")
+        raise ResourceNotFoundException("Examinations")
     return e
 
 
-def create_examination(examination):
-    examination_repository.create_examination(examination)
+def add_course(name, description):
+    existing_course = examination_repository.find_course_by_name(name)
+    if existing_course:
+        raise DuplicateResourceException(name, existing_course.id)
+
+    course = examination_repository.create_course()
+    course.name = name
+    course.description = description
+
+    examination_repository.save_course(course)
+
+
+def add_examination(title, description):
+    pass
 
 
 def create_education():
@@ -60,23 +76,41 @@ def update_education():
     pass
 
 
-def update_course():
-    pass
+def update_course(course_id, name, description):
+    course = examination_repository.find_course_by_id(course_id)
+    if course.name != name and \
+            examination_repository.find_course_by_name(name):
+        raise DuplicateResourceException("Course", name)
+    course.name = name
+    course.description = description
+
+    examination_repository.save_course(course)
 
 
 def delete_examination():
     pass
 
 
-def delete_education(education_id):
-    e = examination_repository.find_all_examinations_by_education(education_id)
-    examination_repository.delete_examination(e)
+def count_examinations_by_course(course_id):
+    exams = examination_repository.find_all_examinations_by_course(course_id)
+    return len(exams)
 
-    examination_repository.delete_education(education_id)
+
+def count_examinations_by_education(education_id):
+    exams = examination_repository.\
+        find_all_examinations_by_education(education_id)
+    return len(exams)
+
+
+def delete_education(education_id):
+    if count_examinations_by_course(education_id) >= 1:
+        raise BusinessRuleException("Education has examinations")
+    else:
+        examination_repository.delete_course(education_id)
 
 
 def delete_course(course_id):
-    exams = examination_repository.find_all_examinations_by_course(course_id)
-    examination_repository.delete_examination(exams)
-
-    examination_repository.delete_course(course_id)
+    if count_examinations_by_course(course_id) >= 1:
+        raise BusinessRuleException("Course has examinations")
+    else:
+        examination_repository.delete_course(course_id)
